refactor: enhance ChangePassword component with validation and alert handling
